+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Unreleased
 
 - [FEATURE] Add support for UK1 Datadog Site. See [#3087][]
+- [FIX] Avoid delaying `DatadogFlags` provider initialization on the initial cached flags read. See [#3078][]
 
 # 3.14.0 / 15-07-2026
 
@@ -1200,6 +1201,7 @@ Release `2.0` introduces breaking changes. Follow the [Migration Guide](MIGRATIO
 [#3019]: https://github.com/DataDog/dd-sdk-ios/pull/3019
 [#3051]: https://github.com/DataDog/dd-sdk-ios/pull/3051
 [#3087]: https://github.com/DataDog/dd-sdk-ios/pull/3087
+[#3078]: https://github.com/DataDog/dd-sdk-ios/pull/3078
 
 [@00fa9a]: https://github.com/00FA9A
 [@britton-earnin]: https://github.com/Britton-Earnin